Add article buttons on bottom if too long
When reading a very long article we want to show navigational buttons on
the bottom as well as on the top, but for short articles we only want
the top ones.

Using https://github.com/customd/jquery-visible we can check when the
browser starts up (reloaded page) or scrolls if the top buttons are
visible. Once they aren't, display the bottom ones.